# docs/SetupDocs/nextxr_openapi.yaml
openapi: 3.0.3
info:
  title: NextXR Digital Twin Platform API
  description: |
    Ontology-driven industrial digital-twin API. Covers twin lifecycle, the
    semantic graph and its governance, the machine-twin physics runtime,
    telemetry ingestion and history, field-protocol connectors, solar PV
    analytics, the embedded AI copilot, agentic twin authoring, 3-D
    reconstruction, the live event bus and platform health.

    136 endpoints across 12 routers plus a mounted 3-D sub-application.

    ## Authentication

    Two credential types, deliberately not interchangeable.

    - **`X-API-Key`** — a human/service credential carrying a **role**
      (`admin`, `write`, `read`) and a **tenant scope**. Configured as a JSON
      array in `NXR_API_KEYS`; the built-in demo keys load **only** when that
      variable is unset (i.e. never in a correctly-configured deployment).
      Set `NXR_REQUIRE_AUTH=1` to reject keyless calls even before keys load.
    - **`X-Device-Token`** — a telemetry device credential granting exactly one
      verb (append telemetry) in exactly one tenant, optionally narrowed to an
      asset prefix. Returned once at registration; only the hash is stored.
      It is never accepted on a read, control or copilot route, and an API key
      is never accepted on the ingest routes in its place.

    > While `NXR_API_KEYS` is unset the API is **OPEN** — every `/api` call,
    > including the LLM-billing `/copilot` endpoints, is served without a key.

...

...
servers:
...
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/entities:
    get:
      tags: [graph]
      summary: List entities for a tenant
      operationId: listEntities
      parameters:
        - $ref: '#/components/parameters/TenantQuery'
        - name: label
          in: query
          schema: { type: string }
          description: Restrict to one taxonomy category (e.g. PhysicalAsset)
        - name: limit
          in: query
          schema: { type: integer, default: 100, maximum: 1000 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Entities in the caller's tenant scope
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Entity' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '503': { $ref: '#/components/responses/GraphUnavailable' }
    post:
      tags: [write]
      summary: Create an entity (SHACL-validated before commit)
      operationId: createEntity
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EntityCreate' }
      responses:
        '201':
          description: Created; a change-log event and a bus event were emitted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Entity' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503': { $ref: '#/components/responses/GraphUnavailable' }
  /api/v1/entities/{node_id}:
    parameters:
      - $ref: '#/components/parameters/NodeId'
    get:
      tags: [graph]
      summary: Get one entity with its properties
      operationId: getEntity
      parameters: [ { $ref: '#/components/parameters/TenantQuery' } ]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Entity' }
          description: The entity
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [write]
      summary: Update entity properties
      operationId: updateEntity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant, properties]
              properties:
                tenant: { type: string }
                properties: { type: object, additionalProperties: true }
      responses:
        '200':
          description: Updated; the change-log diff records old and new per field
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Entity' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
    delete:
      tags: [write]
      summary: Delete an entity
      operationId: deleteEntity
      parameters: [ { $ref: '#/components/parameters/TenantQuery' } ]
      responses:
        '200':
          description: Deleted; a delete event extends the tenant's chain
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ack' }
  /api/v1/entities/{node_id}/rel:
    post:
      tags: [write]
      summary: Relate two entities with an ontology object property
      operationId: relateEntity
      parameters: [ { $ref: '#/components/parameters/NodeId' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant, predicate, target_id]
              properties:
                tenant: { type: string }
                predicate:
                  type: string
                  description: CURIE resolved from the registered prefix table
                  example: hvac:servesSpace
                target_id: { type: string }
      responses:
        '201':
          description: Relationship created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ack' }
        '400':
          description: Unregistered CURIE prefix
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
  /api/v1/schema/class/{name}:
    get:
      tags: [schema]
      summary: Class definition and metadata
      operationId: schemaClass
      parameters: [ { $ref: '#/components/parameters/ClassName' } ]
      responses:
        '200': { $ref: '#/components/responses/JsonObject' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/v1/schema/class/{name}/properties:
    get:
      tags: [schema]
      summary: Observable and object properties of a class
      operationId: schemaClassProperties
      parameters: [ { $ref: '#/components/parameters/ClassName' } ]
      responses:
        '200': { $ref: '#/components/responses/JsonObject' }
...
  /api/v1/schema/validate:
    post:
      tags: [schema]
      summary: SHACL-validate a payload without writing it
      operationId: schemaValidate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [canonical_type, properties]
              properties:
                canonical_type: { type: string }
                properties: { type: object, additionalProperties: true }
      responses:
        '200':
          description: Conformance report
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ValidationReport' }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Carries a role (`admin` / `write` / `read`) and a tenant scope. Configured
        via `NXR_API_KEYS`.
    DeviceToken:
      type: apiKey
      in: header
      name: X-Device-Token
      description: |
        A telemetry device credential: one verb (append), one tenant, optionally
        one asset prefix. Returned once at registration; only the hash is stored.
  parameters:
    TenantQuery:
      name: tenant
      in: query
      required: true
      schema: { type: string }
      description: Tenant id; must be inside the caller's key scope
...
    NodeId:
      name: node_id
      in: path
      required: true
      schema: { type: string }
    ClassName:
      name: name
      in: path
      required: true
      schema: { type: string }
      description: Ontology class local name (e.g. AirHandler)
...
  responses:
    JsonObject:
      description: Success
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
...
    Forbidden:
      description: Tenant outside the caller's key scope, or role insufficient
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationFailed:
      description: SHACL validation failed; nothing was written
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ValidationReport' }
    GraphUnavailable:
      description: |
        Neo4j is unreachable. Converted once, globally, from the driver's
        connection errors so every graph-backed route degrades the same way.
      content:
        application/json:
          schema:
            type: object
            properties:
              detail: { type: string }
              neo4j: { type: string, example: unreachable }
  schemas:
    Ack:
      type: object
      properties:
        ok: { type: boolean }
        detail: { type: string }
    Error:
      type: object
      properties:
        detail: { type: string }
    Entity:
      type: object
      properties:
        id: { type: string }
        tenantId: { type: string }
        label:
          type: string
          description: The taxonomy category used as the Neo4j label
          enum: [PhysicalAsset, MobileAsset, Actor, Location, Process, Observation, Finding, Incident, Document, Capability]
        canonicalType:
          type: string
          description: The OWL class IRI
          example: https://ontology.nextxr.io/v3/solar#PVArray
        displayName: { type: string }
        status: { type: string }
        updatedAt: { type: string, format: date-time }
        properties:
          type: object
          additionalProperties: true
          description: Class properties plus live signal values written by the runtime
    EntityCreate:
      type: object
      required: [tenant, canonical_type, properties]
      properties:
        tenant: { type: string }
        canonical_type: { type: string }
        properties: { type: object, additionalProperties: true }
...
    ValidationReport:
      type: object
      properties:
        conforms: { type: boolean }
        violations:
          type: array
          items:
            type: object
            properties:
              shape: { type: string }
              focus_node: { type: string }
              path: { type: string }
              message: { type: string }
              severity: { type: string }